Move scheduler to ui
