+++ js/bundles/modules/parts/widgets-renovation.js
@@ -8,6 +8,6 @@ renovation.dxrButton = require('../../../renovation/ui/button.j').default;
 renovation.dxrWidget = require('../../../renovation/ui/common/widget.j').default;
 renovation.dxrGridPager = require('../../../renovation/pager/grid_pager.j').default;
 renovation.dxrDataGrid = require('../../../renovation/data_grid/data_grid.j').default;
-renovation.dxrTooltipItemLayout = require('../../../renovation/scheduler/appointment_tooltip/item_layout.j').default;
+renovation.dxrTooltipItemLayout = require('../../../renovation/ui/scheduler/appointment_tooltip/item_layout.j').default;
 /// BUNDLER_PARTS_END
 module.exports = renovation;
