+++ js/renovation/ui/scheduler/appointment_tooltip/__tests__/item_content.test.tsx
