+++ js/renovation/ui/scheduler/appointment_tooltip/__tests__/item_layout.test.tsx
@@ -3,13 +3,13 @@ import { shallow } from 'enzyme';
 import { Fragment } from 'devextreme-generator/component_declaration/common';
 // https://github.com/benmosher/eslint-plugin-import/issues/1699
 /* eslint-disable-next-line import/named */
-import { dxSchedulerAppointment } from '../../../../ui/scheduler';
+import { dxSchedulerAppointment } from '../../../../../ui/scheduler';
 import {
   TooltipItemLayout,
   viewFunction as TooltipItemLayoutView,
   TooltipItemLayoutProps,
 } from '../item_layout';
-import { Button as DeleteButton } from '../../../ui/button';
+import { Button as DeleteButton } from '../../../button';
 import { Marker } from '../marker';
 import { TooltipItemContent } from '../item_content';
 import getCurrentAppointment from '../utils/get_current_appointment';
@@ -18,7 +18,7 @@ jest.mock('../utils/get_current_appointment', () => jest.fn(() => ({
   text: 'currentAppointment',
 })));
 
-jest.mock('../../../ui/button', () => ({ __esModule: true, Button: () => null }));
+jest.mock('../../../button', () => ({ __esModule: true, Button: () => null }));
 jest.mock('../marker', () => ({ __esModule: true, Marker: () => null }));
 jest.mock('../item_content', () => ({ __esModule: true, TooltipItemContent: () => null }));
