+++ js/renovation/ui/scheduler/appointment_tooltip/item_content.tsx
