+++ js/renovation/ui/scheduler/appointment_tooltip/item_layout.tsx
@@ -2,15 +2,15 @@ import {
   Component, ComponentBindings, JSXComponent,
   OneWay, Template, Fragment, Event,
 } from 'devextreme-generator/component_declaration/common';
-import BaseComponent from '../../preact_wrapper/tooltip_item_content';
-import noop from '../../utils/noop';
+import BaseComponent from '../../../preact_wrapper/tooltip_item_content';
+import noop from '../../../utils/noop';
 /* eslint-disable-next-line import/named */
-import { dxSchedulerAppointment } from '../../../ui/scheduler';
+import { dxSchedulerAppointment } from '../../../../ui/scheduler';
 import {
   AppointmentItem, FormattedContent, GetTextAndFormatDateFn, CheckAndDeleteAppointmentFn,
 } from './types';
 import { Marker } from './marker';
-import { Button } from '../../ui/button';
+import { Button } from '../../button';
 import { TooltipItemContent } from './item_content';
 import getCurrentAppointment from './utils/get_current_appointment';
 import { defaultGetTextAndFormatDate } from './utils/default_functions';
