+++ js/renovation/ui/scheduler/appointment_tooltip/types.tsx
@@ -1,8 +1,8 @@
 // https://github.com/benmosher/eslint-plugin-import/issues/1699
 // eslint-disable-next-line import/named
-import { dxSchedulerAppointment } from '../../../ui/scheduler';
+import { dxSchedulerAppointment } from '../../../../ui/scheduler';
 /* eslint-disable-next-line import/named */
-import { dxPromise } from '../../../core/utils/deferred';
+import { dxPromise } from '../../../../core/utils/deferred';
 
 export type Color = string | undefined;
 export type DeferredColor = dxPromise<Color>;
