+++ js/renovation/ui/scheduler/appointment_tooltip/utils/__tests__/default_functions.test.tsx
