+++ js/renovation/ui/scheduler/appointment_tooltip/utils/__tests__/get_current_appointment.test.tsx
