+++ js/renovation/ui/scheduler/appointment_tooltip/utils/get_current_appointment.tsx
@@ -1,6 +1,6 @@
 import { AppointmentItem } from '../types';
 /* eslint-disable-next-line import/named */
-import { dxSchedulerAppointment } from '../../../../ui/scheduler';
+import { dxSchedulerAppointment } from '../../../../../ui/scheduler';
 
 export default (appointmentItem: AppointmentItem): dxSchedulerAppointment => {
   const { settings, data, currentData } = appointmentItem!;
