+++ js/renovation/ui/scheduler/workspaces/base/group_panel/vertical/__tests__/row.test.tsx
