+++ js/renovation/ui/scheduler/workspaces/base/group_panel/vertical/layout.tsx
