+++ js/renovation/ui/scheduler/workspaces/base/time_panel/__tests__/layout.test.tsx
