+++ js/renovation/ui/scheduler/workspaces/day/date_table/__tests__/cell.test.tsx
