+++ js/renovation/ui/scheduler/workspaces/day/date_table/__tests__/layout.test.tsx
