+++ js/renovation/ui/scheduler/workspaces/month/date_table/__tests__/cell.test.tsx
