+++ js/renovation/ui/scheduler/workspaces/month/date_table/cell.tsx
