+++ js/renovation/ui/scheduler/workspaces/month/header_panel/__tests__/cell.test.tsx
@@ -3,9 +3,9 @@ import {
   MonthHeaderPanelCell as Cell,
   viewFunction as CellView,
 } from '../cell';
-import dateLocalization from '../../../../../../localization/date';
+import dateLocalization from '../../../../../../../localization/date';
 
-jest.mock('../../../../../../localization/date', () => ({
+jest.mock('../../../../../../../localization/date', () => ({
   getDayNames: jest.fn(() => [0, 1, 2, 3, 4, 5, 6, 7]),
 }));
