+++ js/renovation/ui/scheduler/workspaces/month/header_panel/cell.tsx
@@ -1,7 +1,7 @@
 import {
   Component, ComponentBindings, JSXComponent, OneWay,
 } from 'devextreme-generator/component_declaration/common';
-import dateLocalization from '../../../../../localization/date';
+import dateLocalization from '../../../../../../localization/date';
 
 export const viewFunction = (viewModel: MonthHeaderPanelCell) => (
   <td
